Restore feedback files
